Restructure utils
