+++ grazer/util.py
@@ -1,4 +1,11 @@
+import requests
+import logging
+
 from six.moves import zip_longest
+from copy import copy
+from functools import reduce
+
+logger = logging.getLogger("Util")
 
 
 def time_convert(span):
@@ -17,3 +24,42 @@ def time_convert(span):
 def grouper(n, iterable):
     args = [iter(iterable)] * n
     return zip_longest(fillvalue=None, *args)
+
+
+def get_session(cookies=None):
+    session = requests.Session()
+    if cookies:
+        requests.utils.add_dict_to_cookiejar(session.cookies,
+                                             cookies)
+    return session
+
+
+def extract_links(page, ignore_hashes=True):
+    gen = [a.get("href") for a in page.find_all("a")]
+    if ignore_hashes:
+        return list(set(map(lambda x: x.split("#")[0],
+                        filter(lambda x: x is not None, gen))))
+    else:
+        return list(set(gen))
+
+
+def trim_link(link, domain):
+    if link is None or len(link) == 0:
+        return None
+
+    if not link.startswith("http"):
+        return ("/" if link[0] != "/" else "") + link
+
+    if domain not in link:
+        # External link to another domain
+        logger.debug("Link {0} is not in {1} domain".format(link, domain))
+        return None
+
+    prefixes = ["https://", "http://", "://"]
+    original_link = copy(link)
+    link = reduce(lambda a, b: a.replace(b, ""), prefixes, link)
+    if "/" in link:
+        start, end = link.split("/", 1)
+        return "/" + end
+    else:
+        return original_link
